Reuse Lombscargle plan in LS
We are reusing the lombscargle plan in the simulated annealing
algorithm.
This works, because we are only shuffling the data around and we can
instead shuffle the time vector and get the same result.
This is the same trick as in the LombScargle.bootstrap function.
At the end we use the permutation of the time vector to permute the
input signal to get a surrogate vector.